backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "aiosqlite>=0.22.1",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.8",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "pydantic-settings>=2.11.0",
    "sqlmodel>=0.0.34",
    "uvicorn[standard]>=0.39.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pyinstaller>=6.19.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
]
